# docs/swagger.yaml
basePath: /api/
definitions:
  api.ParamSearch:
    properties:
      filter:
        additionalProperties: true
        type: object
      keyword:
        additionalProperties:
          type: string
        type: object
      limit:
        type: integer
      skip:
        type: integer
      sort:
        additionalProperties:
          type: integer
        type: object
    type: object
...
  api.ReplyData-int64:
    properties:
      data:
        type: integer
      error:
        type: string
    type: object
...
  api.ReplyData-types_Serial:
    properties:
      data:
        $ref: '#/definitions/types.Serial'
      error:
        type: string
    type: object
  api.ReplyData-types_Server:
    properties:
      data:
        $ref: '#/definitions/types.Server'
      error:
        type: string
    type: object
...
  api.ReplyList-types_Serial:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Serial'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
  api.ReplyList-types_Server:
    properties:
      data:
        items:
          $ref: '#/definitions/types.Server'
        type: array
      error:
        type: string
      total:
        type: integer
    type: object
...
  types.DefaultDevice:
    properties:
      name:
        description: 名称
        type: string
      product_id:
        description: 产品ID
        type: string
      slave:
        description: 从站号
        type: integer
    type: object
...
  types.Serial:
    properties:
      baud_rate:
        description: 9600 115200
        type: integer
      created:
        description: 创建时间
        type: string
      data_bits:
        description: 5 6 7 8
        type: integer
      desc:
        description: 说明
        type: string
      disabled:
        type: boolean
      heartbeat:
        description: 心跳包
        type: string
      id:
        description: ID
        type: string
      name:
        description: 名称
        type: string
      parity_mode:
        description: 0 1 2 NONE ODD EVEN
        type: integer
      poller_interval:
        description: 采集间隔
        type: integer
      poller_period:
        description: 采集周期
        type: integer
      port_name:
        description: port, e.g. COM1 "/dev/ttySerial1".
        type: string
      protocol_name:
        description: 协议 rtu tcp parallel-tcp
        type: string
      protocol_options:
        description: 协议参数
        type: string
      retry_maximum:
        description: 最大次数
        type: integer
      retry_timeout:
        description: 重试时间
        type: integer
      stop_bits:
        description: 1 2
        type: integer
    type: object
  types.Server:
    properties:
      created:
        description: 创建时间
        type: string
      desc:
        description: 说明
        type: string
      devices:
        items:
          $ref: '#/definitions/types.DefaultDevice'
        type: array
      disabled:
        type: boolean
      heartbeat:
        description: 心跳包
        type: string
      id:
        description: ID
        type: string
      name:
        description: 名称
        type: string
      poller_interval:
        description: 采集间隔
        type: integer
      poller_period:
        description: 采集周期
        type: integer
      port:
        description: 监听端口
        type: integer
      protocol_name:
        description: 协议 rtu tcp parallel-tcp
        type: string
      protocol_options:
        description: 协议参数
        type: string
      standalone:
        description: 单例模式（不支持注册）
        type: boolean
    type: object
info:
  contact: {}
  description: API文档
  title: 物联大师网关接口文档
  version: 1.0 版本
paths:
...
  /serial/{id}:
    get:
      consumes:
      - application/json
      description: 获取串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Serial'
      summary: 获取串口
      tags:
      - serial
    post:
      consumes:
      - application/json
      description: 修改串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      - description: 串口信息
        in: body
        name: serial
        required: true
        schema:
          $ref: '#/definitions/types.Serial'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Serial'
      summary: 修改串口
      tags:
      - serial
  /serial/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Serial'
      summary: 删除串口
      tags:
      - serial
  /serial/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Serial'
      summary: 禁用串口
      tags:
      - serial
  /serial/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用串口
      parameters:
      - description: 串口ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Serial'
      summary: 启用串口
      tags:
      - serial
  /serial/count:
    post:
      consumes:
      - application/json
      description: 查询串口数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 查询串口数量
      tags:
      - serial
  /serial/create:
    post:
      consumes:
      - application/json
      description: 创建串口
      parameters:
      - description: 串口信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Serial'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Serial'
      summary: 创建串口
      tags:
      - serial
...
  /serial/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入串口
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 导入串口
      tags:
      - product
  /serial/list:
    get:
      consumes:
      - application/json
      description: 查询串口
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Serial'
      summary: 查询串口
      tags:
      - serial
...
  /serial/search:
    post:
      consumes:
      - application/json
      description: 查询串口
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Serial'
      summary: 查询串口
      tags:
      - serial
  /server/{id}:
    get:
      consumes:
      - application/json
      description: 获取服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Server'
      summary: 获取服务器
      tags:
      - server
    post:
      consumes:
      - application/json
      description: 修改服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      - description: 服务器信息
        in: body
        name: server
        required: true
        schema:
          $ref: '#/definitions/types.Server'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Server'
      summary: 修改服务器
      tags:
      - server
  /server/{id}/delete:
    get:
      consumes:
      - application/json
      description: 删除服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Server'
      summary: 删除服务器
      tags:
      - server
  /server/{id}/disable:
    get:
      consumes:
      - application/json
      description: 禁用服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Server'
      summary: 禁用服务器
      tags:
      - server
  /server/{id}/enable:
    get:
      consumes:
      - application/json
      description: 启用服务器
      parameters:
      - description: 服务器ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Server'
      summary: 启用服务器
      tags:
      - server
  /server/count:
    post:
      consumes:
      - application/json
      description: 查询服务器数量
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 查询服务器数量
      tags:
      - server
  /server/create:
    post:
      consumes:
      - application/json
      description: 创建服务器
      parameters:
      - description: 服务器信息
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/types.Server'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-types_Server'
      summary: 创建服务器
      tags:
      - server
...
  /server/import:
    post:
      consumes:
      - multipart/form-data
      description: 导入服务端
      parameters:
      - description: 压缩包
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyData-int64'
      summary: 导入服务端
      tags:
      - product
  /server/list:
    get:
      consumes:
      - application/json
      description: 查询服务器
      parameters:
      - in: query
        name: limit
        type: integer
      - in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Server'
      summary: 查询服务器
      tags:
      - server
  /server/search:
    post:
      consumes:
      - application/json
      description: 查询服务器
      parameters:
      - description: 查询参数
        in: body
        name: search
        required: true
        schema:
          $ref: '#/definitions/api.ParamSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ReplyList-types_Server'
      summary: 查询服务器
      tags:
      - server
swagger: "2.0"